requirements.txt.txt
python-dateutil
Django
djangorestframework